requirements.txt
django>=4.2,<6.0
waitress>=2.1
flask>=3.0
pywin32>=306
requests>=2.31
python-dateutil>=2.8
pillow
pywin32
arabic-reshaper
python-bidi